git-svn-id: https://fffunctions.googlecode.com/svn/trunk@49 cb06e305-f834-26d0-b1f6-d1fa8a6c78bd

+++ roxygen.R
@@ -9,4 +9,6 @@ roxygenize( 'pkg'
           , use.Rd2 = TRUE
 		    )
 
-unlink( 'pkg/inst/doc', TRUE)
+if (length(list.files('pkg/inst/doc')) == 0){
+   unlink( 'pkg/inst/doc', TRUE)   
+}
